+++ kwarray/__init__.py
@@ -20,7 +20,7 @@ __explicit__ = [
     'ArrayAPI', 'dtype_info',
 ]
 
-__version__ = '0.5.18'
+__version__ = '0.5.19'
 
 from kwarray.arrayapi import ArrayAPI, dtype_info