+++ kwarray/util_slider.py
@@ -84,25 +84,24 @@ class SlidingWindow(ub.NiceRepr):
          [(4, 0), (4, 1), (4, 2), (4, 3), (4, 4)],
          [(5, 0), (5, 1), (5, 2), (5, 3), (5, 4)],
          [(6, 0), (6, 1), (6, 2), (6, 3), (6, 4)]]
+
+    Example:
+        >>> # Test shapes that dont fit
+        >>> # When the window is bigger than the shape, the left-aligned slices
+        >>> # are returend.
+        >>> self = SlidingWindow((3, 3), (12, 12), allow_overshoot=True, keepbound=True)
+        >>> print(list(self))
+        [(slice(0, 12, None), slice(0, 12, None))]
+        >>> print(list(SlidingWindow((3, 3), None, allow_overshoot=True, keepbound=True)))
+        [(slice(0, 3, None), slice(0, 3, None))]
+        >>> print(list(SlidingWindow((3, 3), (None, 2), allow_overshoot=True, keepbound=True)))
+        [(slice(0, 3, None), slice(0, 2, None)), (slice(0, 3, None), slice(1, 3, None))]
     """
     def __init__(self, shape, window, overlap=None, stride=None,
                  keepbound=False, allow_overshoot=False):
-        if overlap is None and stride is None:
-            overlap = 0
 
-        assert len(window) == len(shape), (
-            'incompatible dims: {} {}'.format(len(window),
-                                              len(shape)))
-        # assert all(d <= D for d, D in zip(window, shape)), (
-        #         'window must be smaller than shape')
-
-        stride, overlap = self._compute_stride(overlap, stride, shape,
-                                               window)
-
-        # print('stride = {!r}'.format(stride))
-        if not all(stride):
-            raise ValueError(
-                'Step must be positive everywhere. Got={}'.format(stride))
+        stride, overlap, window = self._compute_stride(
+            overlap, stride, shape, window)
 
         stide_kw = [dict(margin=d, stop=D, step=s, keepbound=keepbound,
                          check=not keepbound and not allow_overshoot)
@@ -154,10 +153,24 @@ class SlidingWindow(ub.NiceRepr):
         Ensures that stride hasoverlap the correct shape.  If stride is not
         provided, compute stride from desired overlap.
         """
+        if window is None:
+            window = shape
+
         if isinstance(stride, np.ndarray):
             stride = tuple(stride)
+
         if isinstance(overlap, np.ndarray):
             overlap = tuple(overlap)
+
+        if len(window) != len(shape):
+            raise ValueError('incompatible dims: {} {}'.format(len(window),
+                                                               len(shape)))
+
+        if any(d is None for d in window):
+            window = [D if d is None else d for d, D in zip(window, shape)]
+
+        if overlap is None and stride is None:
+            overlap = 0
 
         if not (overlap is None) ^ (stride is None):
             raise ValueError('specify overlap({}) XOR stride ({})'.format(
@@ -176,8 +189,13 @@ class SlidingWindow(ub.NiceRepr):
                 stride = [stride] * len(window)
         # Recompute fractional overlap after integer stride is computed
         overlap = [(d - s) / d for s, d in zip(stride, window)]
+
         assert len(stride) == len(shape), 'incompatible dims'
-        return stride, overlap
+
+        if not all(stride):
+            raise ValueError(
+                'Step must be positive everywhere. Got={}'.format(stride))
+        return stride, overlap, window
 
     def __len__(self):
         return self.n_total
@@ -387,13 +405,15 @@ def _slices1d(margin, stop, step=None, start=0, keepbound=False, check=True):
     Helper to generates slices in a single dimension.
 
     Args:
-        start (int): starting point (in most cases set this to 0)
 
         margin (int): the length of the slice (window)
 
         stop (int): the length of the image dimension
 
-        step (int): the length of each step / distance between slices
+        step (int, default=None): the length of each step / distance between
+            slices
+
+        start (int, default=0): starting point (in most cases set this to 0)
 
         keepbound (bool): if True, a non-uniform step will be taken to ensure
             that the right / bottom of the image is returned as a slice if
@@ -455,5 +475,7 @@ def _slices1d(margin, stop, step=None, start=0, keepbound=False, check=True):
                 # Ensure the boundary is always used even if steps
                 # would overshoot Could do some other strategy here
                 pos = stop - margin
+                if pos < 0:
+                    break
             else:
                 break